Moved event attribute containers to a separate subdirectory.
