+++ tests/analysis/tagging.py
@@ -7,18 +7,18 @@ import unittest
 from plaso.analysis import tagging
 from plaso.engine import queue
 from plaso.engine import single_process
-from plaso.lib import event
 from plaso.lib import timelib
+from plaso.containers import events
 
 from tests.analysis import test_lib
 
 
-class TestPrefetchEvent(event.EventObject):
+class TestPrefetchEvent(events.EventObject):
   """A test event type for the tagging analysis plugin."""
   DATA_TYPE = u'windows:prefetch'
 
 
-class TestChromeDownloadEvent(event.EventObject):
+class TestChromeDownloadEvent(events.EventObject):
   """A test event type for the tagging analysis plugin."""
   DATA_TYPE = u'chrome:history:file_downloaded'
 
@@ -28,7 +28,7 @@ class TaggingTest(test_lib.AnalysisPluginTestCase):
   _INVALID_TEST_TAG_FILE_NAME = u'invalid_test_tag_file.txt'
   _TEST_TAG_FILE_NAME = u'test_tag_file.txt'
 
-  _TEST_EVENTS = [
+  _EVENT_DICTS = [
       {u'event_type': u'prefetch',
        u'timestamp': timelib.Timestamp.CopyFromString(u'2015-05-01 15:12:00'),
        u'attributes': {}
@@ -50,7 +50,7 @@ class TaggingTest(test_lib.AnalysisPluginTestCase):
     elif test_event_dict[u'event_type'] == u'chrome_download':
       event_object = TestChromeDownloadEvent()
     else:
-      event_object = event.EventObject()
+      event_object = events.EventObject()
 
     event_object.timestamp = test_event_dict[u'timestamp']
     for key, value in test_event_dict[u'attributes']:
@@ -85,9 +85,10 @@ class TaggingTest(test_lib.AnalysisPluginTestCase):
     """Test that the tagging plugin successfully tags events."""
     event_queue = single_process.SingleProcessQueue()
     test_queue_producer = queue.ItemQueueProducer(event_queue)
-    events = [self._CreateTestEventObject(test_event)
-              for test_event in self._TEST_EVENTS]
-    test_queue_producer.ProduceItems(events)
+    event_objects = [
+        self._CreateTestEventObject(test_event)
+        for test_event in self._EVENT_DICTS]
+    test_queue_producer.ProduceItems(event_objects)
     analysis_plugin = tagging.TaggingPlugin(event_queue)
     test_file = self._GetTestFilePath([self._TEST_TAG_FILE_NAME])
     analysis_plugin.SetAndLoadTagFile(test_file)
