+++ tests/serializer/protobuf_serializer.py
@@ -5,6 +5,7 @@
 import collections
 import unittest
 
+from plaso.containers import events
 from plaso.lib import event
 from plaso.proto import plaso_storage_pb2
 from plaso.serializer import protobuf_serializer
@@ -188,7 +189,7 @@ class ProtobufEventObjectSerializerTest(ProtobufSerializerTestCase):
 
   def testWriteSerialized(self):
     """Tests the WriteSerialized function."""
-    event_object = event.EventObject()
+    event_object = events.EventObject()
 
     event_object.data_type = u'test:event2'
     event_object.timestamp = 1234124
@@ -225,7 +226,7 @@ class ProtobufEventTagSerializerTest(ProtobufSerializerTestCase):
 
   def setUp(self):
     """Makes preparations before running an individual test."""
-    self._event_tag = event.EventTag(comment=u'My first comment.')
+    self._event_tag = events.EventTag(comment=u'My first comment.')
     self._event_tag.store_number = 234
     self._event_tag.store_index = 18
     self._event_tag.AddLabels([u'Malware', u'Common'])
